# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Minimal backend framework APIs (users, auth, options, payment, order, status).",
        "title": "Done Hub Minimal API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/order/": {
            "get": {
                "description": "查询订单列表（管理员）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "List orders (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "创建充值/支付订单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "description": "订单请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.OrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/order/status": {
            "get": {
                "description": "查询订单状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Check order status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单号",
                        "name": "trade_no",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/payment/": {
            "get": {
                "description": "获取支付网关列表（管理员）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "List payments (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "更新支付网关（管理员）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Update payment (admin)",
                "parameters": [
                    {
                        "description": "网关信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Payment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "新增支付网关（管理员）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Create payment (admin)",
                "parameters": [
                    {
                        "description": "网关信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Payment"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/payment/notify/{uuid}": {
            "post": {
                "description": "支付回调入口",
                "tags": [
                    "Payment"
                ],
                "summary": "Payment callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "回调网关UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/payment/{id}": {
            "get": {
                "description": "获取支付网关详情（管理员）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Get payment (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "网关ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "删除支付网关（管理员）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Delete payment (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "网关ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.OrderRequest": {
            "type": "object",
            "required": [
                "amount",
                "uuid"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "model.CurrencyType": {
            "type": "string",
            "enum": [
                "USD",
                "CNY"
            ],
            "x-enum-varnames": [
                "CurrencyTypeUSD",
                "CurrencyTypeCNY"
            ]
        },
...
        "model.Payment": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "Config 网关配置（JSON 文本，按网关类型定义）",
                    "type": "string"
                },
                "created_at": {
                    "description": "CreatedAt 创建时间（Unix 秒）",
                    "type": "integer"
                },
                "currency": {
                    "description": "Currency 币种（USD/CNY）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.CurrencyType"
                        }
                    ]
                },
                "enable": {
                    "description": "Enable 是否启用",
                    "type": "boolean"
                },
                "fixed_fee": {
                    "description": "FixedFee 固定手续费",
                    "type": "number"
                },
                "icon": {
                    "description": "Icon 图标URL（前端展示）",
                    "type": "string"
                },
                "id": {
                    "description": "ID 主键自增ID",
                    "type": "integer"
                },
                "name": {
                    "description": "Name 网关名称（展示使用）",
                    "type": "string"
                },
                "notify_domain": {
                    "description": "NotifyDomain 回调域名（如需外网回调，配置对外可访问域名）",
                    "type": "string"
                },
                "percent_fee": {
                    "description": "PercentFee 百分比手续费（0~1，示例：0.05=5%）",
                    "type": "number"
                },
                "sort": {
                    "description": "Sort 排序（数字越小越靠前）",
                    "type": "integer"
                },
                "type": {
                    "description": "Type 网关类型（如 stripe/alipay/wxpay/epay 等）",
                    "type": "string"
                },
                "uuid": {
                    "description": "UUID 网关唯一标识（用于回调识别），系统生成",
                    "type": "string"
                }
            }
        },
...
    }
}